requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.24
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
email-validator==2.1.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pypdf==3.17.4 # Replaced PyPDF2 with pypdf
python-docx==1.1.0
gpt4all
Pillow>=10.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
authlib==1.2.1
pytesseract==0.3.10
pandas==2.1.4 # Added explicitly to ensure correct installation with numpy
numpy>=1.24.0
openai-whisper
opencv-python
pytest-asyncio
pytest
